requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
email-validator==2.2.0
pytest==8.2.2
black
ruff
pre-commit
